requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

streamlit==1.38.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pillow==10.4.0
